src/backend/server/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.0
aiohttp==3.9.0
websockets==12.0
fastapi==0.115.5
uvicorn[standard]==0.32.0
pytest==7.4.0
pytest-asyncio==0.21.0
